requirements.txt
Flask
SQLAlchemy
alembic
pydantic
python-dateutil
pytest
